+++ packages/atomic/src/utils/custom-element-tags.ts
@@ -61,6 +61,7 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-icon',
   'atomic-insight-edit-toggle',
   'atomic-insight-facet',
+  'atomic-insight-folded-result-list',
   'atomic-insight-full-search-button',
   'atomic-insight-generate-answer-button',
   'atomic-insight-history-toggle',
@@ -83,6 +84,7 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-insight-search-box',
   'atomic-insight-smart-snippet',
   'atomic-insight-smart-snippet-feedback-modal',
+  'atomic-insight-smart-snippet-suggestions',
   'atomic-insight-tab',
   'atomic-insight-tabs',
   'atomic-insight-user-actions-modal',
@@ -135,6 +137,7 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-rating-range-facet',
   'atomic-recs-error',
   'atomic-recs-interface',
+  'atomic-recs-list',
   'atomic-recs-result',
   'atomic-recs-result-template',
   'atomic-refine-modal',
@@ -197,6 +200,7 @@ export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set<string>([
   'atomic-text',
   'atomic-timeframe',
   'atomic-timeframe-facet',
+  'generated-answer-thread-item',
 ]);
 
 /**